# axum/openapi.yaml
openapi: 3.1.0
info:
  title: Ogonek API
  description: ''
  contact:
    name: API Support
    email: danila.volkov@noxlovette.com
  license:
    name: ''
  version: 1.6.0
servers:
- url: https://api.ogonek.app
  description: Production server
- url: http://localhost:3000
  description: Local development server
paths:
  /api/v1/auth/bind:
    post:
      tags:
      - Auth
      summary: Binds the student to the teacher
      operationId: bind_student_to_teacher
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindPayload'
        required: true
      responses:
        '204':
          description: Student bound to teacher successfully
        '400':
          description: Invalid bind data
        '401':
          description: Invalid invite token
  /api/v1/auth/invite:
    get:
      tags:
      - Auth
      summary: Generates the invite link for the teacher
      operationId: generate_invite_link
      parameters:
      - name: invite
        in: query
        description: Invite token
        required: true
        schema:
          $ref: '#/components/schemas/InviteQuery'
      responses:
        '200':
          description: Invite link generated
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Invalid invite token
        '401':
          description: Unauthorized
      security:
      - api_key: []
  /api/v1/auth/refresh:
    post:
      tags:
      - Auth
      summary: Receives the refresh token as json, gets it, then decodes, finds the user id, and generates a new access token
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenPayload'
        required: true
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          description: Invalid refresh token
  /api/v1/auth/signin:
    post:
      tags:
      - Auth
      operationId: signin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthPayload'
        required: true
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Invalid credentials
  /api/v1/auth/signup:
    post:
      tags:
      - Auth
      operationId: signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpPayload'
        required: true
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid registration data
        '409':
          description: User already exists
...
  /api/v1/learn:
    get:
      tags:
      - Learn
      summary: Returns the list of all cards due for review
      operationId: fetch_due_cards
      responses:
        '200':
          description: Due cards fetched successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardProgressWithFields'
        '401':
          description: Unauthorized
      security:
      - api_key: []
  /api/v1/learn/subscribe/{id}:
    post:
      tags:
      - Learn
      summary: Subscribes the user to the deck
      operationId: subscribe_to_deck
      parameters:
      - name: id
        in: path
        description: Deck ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deck subscribed to successfully
        '401':
          description: Unauthorized
        '404':
          description: Deck not found
      security:
      - api_key: []
    delete:
      tags:
      - Learn
      summary: Unsubscribes the user from the deck
      operationId: unsubscribe_from_deck
      parameters:
      - name: id
        in: path
        description: Deck ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deck subscribed to successfully
        '401':
          description: Unauthorized
        '404':
          description: Deck not found
      security:
      - api_key: []
  /api/v1/learn/{id}:
    put:
      tags:
      - Learn
      summary: Updates the learn progress on a card
      operationId: update_card_progress
      parameters:
      - name: id
        in: path
        description: Card ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Card progress updated successfully
        '401':
          description: Unauthorized
      security:
      - api_key: []
    delete:
      tags:
      - Learn
      summary: Resets the progress for a particular deck
      operationId: reset_deck_progress
      parameters:
      - name: id
        in: path
        description: Deck ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Card progress updated successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardProgressWithFields'
        '401':
          description: Unauthorized
      security:
      - api_key: []
...
components:
  schemas:
...
    AuthPayload:
      type: object
      required:
      - username
      - pass
      properties:
        pass:
          type: string
        username:
          type: string
...
    BindPayload:
      type: object
      required:
      - studentId
      - inviteToken
      properties:
        inviteToken:
          type: string
        studentId:
          type: string
...
    CardProgressWithFields:
      type: object
      required:
      - id
      - userId
      - cardId
      - reviewCount
      - easeFactor
      - interval
      - front
      - back
      properties:
        back:
          type: string
        cardId:
          type: string
        dueDate:
          type:
          - string
          - 'null'
          format: date-time
        easeFactor:
          type: number
          format: double
        front:
          type: string
        id:
          type: string
        interval:
          type: integer
          format: int32
        lastReviewed:
          type:
          - string
          - 'null'
          format: date-time
        mediaUrl:
          type:
          - string
          - 'null'
        reviewCount:
          type: integer
          format: int32
        userId:
          type: string
...
    InviteQuery:
      type: object
      required:
      - isRegistered
      properties:
        isRegistered:
          type: string
...
    RefreshTokenPayload:
      type: object
      required:
      - refreshToken
      properties:
        refreshToken:
          type: string
    RefreshTokenResponse:
      type: object
      required:
      - accessToken
      properties:
        accessToken:
          $ref: '#/components/schemas/TokenWithExpiry'
    SignUpPayload:
      type: object
      required:
      - name
      - email
      - pass
      - username
      - role
      properties:
        email:
          type: string
        name:
          type: string
        pass:
          type: string
        role:
          type: string
        username:
          type: string
...
    TokenPair:
      type: object
      required:
      - accessToken
      - refreshToken
      properties:
        accessToken:
          $ref: '#/components/schemas/TokenWithExpiry'
        refreshToken:
          $ref: '#/components/schemas/TokenWithExpiry'
    TokenWithExpiry:
      type: object
      required:
      - token
      - expiresAt
      properties:
        expiresAt:
          type: integer
          format: int64
          minimum: 0
        token:
          type: string
...
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
...